# internal/features/tasks/transport/http/patch_task.go
package tasks_transport_http

import (
	"fmt"
	"net/http"

	"github.com/Lockok/golang-niltodo/internal/core/domain"
	core_logger "github.com/Lockok/golang-niltodo/internal/core/logger"
	core_http_request "github.com/Lockok/golang-niltodo/internal/core/transport/http/request"
	core_http_response "github.com/Lockok/golang-niltodo/internal/core/transport/http/response"
	core_http_types "github.com/Lockok/golang-niltodo/internal/core/transport/http/types"
)

type PatchTaskRequest struct {
	Title core_http_types.Nullable[string]  `json:"title" swaggertype:"string" example:"Решить задачу"`
	Description core_http_types.Nullable[string] `json:"description" swaggertype:"description" example:"null"`
	Completed core_http_types.Nullable[bool] `json:"completed" swaggertype:"boolean"`
}

type PatchTaskRequestSwagger struct {
    Title       *string `json:"title" example:"Решить задачу"`
    Description *string `json:"description" example:"Описание"`
    Completed   *bool   `json:"completed" example:"true"`
}

func (r *PatchTaskRequest) Validate() error {
	if r.Title.Set {
		if r.Title.Value == nil {
			return fmt.Errorf("`Title` cant be NULL")
		}

		titleLen := len([]rune(*r.Title.Value))
		if titleLen < 1 || titleLen > 100 {
			return fmt.Errorf("`Title` must be between 1 and 100 symbols")
		}
	}

	if r.Description.Set {
		if r.Description.Value == nil {
			descriptionLen := len([]rune(*r.Description.Value))
			if descriptionLen < 1 || descriptionLen > 1000 {
				return fmt.Errorf("`Description` must be between 1 and 1000 symbols")
			}
		}
	}

	if r.Completed.Set {
		if r.Completed.Value == nil {
			return fmt.Errorf("`Completed` cant be NULL")
		}
	}

	return nil
}

type PatchUserResponse TaskDTOResponse

// PatchTask		godoc
// @Summary			Обновить задачу
// @Description		Обновляет информацию об уже существующей в системе задаче
// @Description		### Логика обновления полей (Three-state logic):
// @Description		1. **Поле не передано**: `description` игнорируется, значение в БД не меняется
// @Description		2. **Явно передано значение**: `"description": "Утром съесть конфету"`
// @Description		3. **Явно передан null**: `"description": null` - очищает поле в БД (set to NULL)
// @Description		Ограничения: `title` и `completed` не могут быть выставлены как NULL
// @Tags			tasks
// @Accept			json
// @Produce			json
// @Param			id 		path     int 					 true  	  "ID изменяемой задачи"
// @Param			request body     PatchTaskRequestSwagger true	  "PatchTask тело запроса"
// @Success			200	 	{object} PatchUserResponse			  	  "Успешно измененная задача"
// @Failure			400		{object} core_http_response.ErrorResponse "Bad request"
// @Failure			404		{object} core_http_response.ErrorResponse "Task not found"
// @Failure			409		{object} core_http_response.ErrorResponse "Conflict"
// @Failure			500		{object} core_http_response.ErrorResponse "Internal server error"
// @Router			/tasks/{id} [patch]
func (h *TasksHTTPHandler) PatchTask(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, rw)

	taskID, err := core_http_request.GetIntPathValue(r, "id")
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to get taskID path value")
	
		return
	}

	var request PatchTaskRequest
	if err := core_http_request.DecodeAndValidateRequest(r, &request); err != nil {
		responseHandler.ErrorResponse(err, "failed to decode and validate HTTP request")
		
		return
	}

	taskPatch := taskPatchFromRequest(request)

	taskDomain, err := h.tasksService.PatchTask(ctx, taskID, taskPatch)
	if err != nil {
		responseHandler.ErrorResponse(
		err,
		"failed to patch task",
		)

		return
	}

	response := PatchUserResponse(taskDTOFromDomain(taskDomain))

	responseHandler.JSONResponse(response, http.StatusOK)
}

 func taskPatchFromRequest(request PatchTaskRequest) domain.TaskPatch {
	return domain.NewTaskPatch(
		request.Title.ToDomain(),
		request.Description.ToDomain(),
		request.Completed.ToDomain(),
	)
}