# internal/features/users/transport/http/create_user.go
package users_transport_http

import (
 	"net/http"

	"github.com/Lockok/golang-niltodo/internal/core/domain"
	core_logger "github.com/Lockok/golang-niltodo/internal/core/logger"
	core_http_request "github.com/Lockok/golang-niltodo/internal/core/transport/http/request"
	core_http_response "github.com/Lockok/golang-niltodo/internal/core/transport/http/response"
)

type CreateUserRequest struct {
	FullName	 string		`json:"full_name" validate:"required,min=3,max=100"							  example:"Vanek Vankov"`
	PhoneNumber	*string		`json:"phone_number" validate:"omitempty,required,min=10,max=15,startswith=+" example:"+79991112233"`
}

type CreateUserResponse UserDTOResponse

// CreateUser		godoc
// @Summary			Создать пользователя
// @Description		Создать нового пользователя в системе
// @Tags			users
// @Accept			json
// @Produce			json
// @Param			request body  	 CreateUserRequest true "CreateUser тело запроса"
// @Success			201		{object} CreateUserResponse "Успешно созданный пользователь"
// @Failure			401		{object} core_http_response.ErrorResponse "Bad request"
// @Failure			500		{object} core_http_response.ErrorResponse "Internal server error"
// @Router			/users	[post]
func (h *UsersHTTPHandler) CreateUser(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, w)

	log.Debug("invoke CreateUser handler")
	
	var request CreateUserRequest

	if err := core_http_request.DecodeAndValidateRequest(r, &request); err != nil {
		responseHandler.ErrorResponse(err, "failed to decode and validate HTTP request")

		return
	}

	userDomain := domainFromDTO(request)

	userDomain, err := h.usersService.CreateUser(ctx, userDomain)
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to create User")														
		
		return
	}

	response := CreateUserResponse(userDtoFromDomain(userDomain))

	responseHandler.JSONResponse(response, http.StatusCreated)
} 

func domainFromDTO(dto CreateUserRequest) domain.User {
	return domain.NewUserUninitialized(dto.FullName, dto.PhoneNumber)
}